+++ modules/express/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.2](https://github.com/BitGo/BitGoJS/compare/@bitgo/express@10.1.1...@bitgo/express@10.1.2) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/express
+
 ## [10.1.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/express@10.1.0...@bitgo/express@10.1.1) (2024-07-24)
 
 **Note:** Version bump only for package @bitgo/express