+++ modules/sdk-coin-bsc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [22.1.12](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-bsc@22.1.11...@bitgo/sdk-coin-bsc@22.1.12) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-bsc
+
 ## [22.1.11](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-bsc@22.1.10...@bitgo/sdk-coin-bsc@22.1.11) (2024-07-24)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-bsc