+++ modules/sdk-coin-btc/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-btc@2.1.2...@bitgo/sdk-coin-btc@2.1.3) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-btc
+
 ## [2.1.2](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-btc@2.1.1...@bitgo/sdk-coin-btc@2.1.2) (2024-07-24)
 
 ### Bug Fixes