+++ modules/sdk-coin-celo/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.20](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-celo@3.0.19...@bitgo/sdk-coin-celo@3.0.20) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-celo
+
 ## [3.0.19](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-celo@3.0.18...@bitgo/sdk-coin-celo@3.0.19) (2024-07-24)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-celo